Move autosave components to utils dir and shorten class names
