+++ spyder/plugins/editor/utils/tests/test_autosave.py
@@ -9,14 +9,14 @@
 import pytest
 
 # Local imports
-from spyder.plugins.editor.autosave import AutosaveComponentForEditorPlugin
+from spyder.plugins.editor.utils.autosave import AutosaveForPlugin
 
 
 def test_autosave_component_set_interval(qtbot, mocker):
     """Test that setting the interval does indeed change it and calls
     do_autosave if enabled."""
-    mocker.patch.object(AutosaveComponentForEditorPlugin, 'do_autosave')
-    addon = AutosaveComponentForEditorPlugin(None)
+    mocker.patch.object(AutosaveForPlugin, 'do_autosave')
+    addon = AutosaveForPlugin(None)
     addon.do_autosave.assert_not_called()
     addon.interval = 10000
     assert addon.interval == 10000
@@ -29,8 +29,8 @@ def test_autosave_component_set_interval(qtbot, mocker):
 @pytest.mark.parametrize('enabled', [False, True])
 def test_autosave_component_timer_if_enabled(qtbot, mocker, enabled):
     """Test that AutosaveCompenent calls do_autosave() on timer if enabled."""
-    mocker.patch.object(AutosaveComponentForEditorPlugin, 'do_autosave')
-    addon = AutosaveComponentForEditorPlugin(None)
+    mocker.patch.object(AutosaveForPlugin, 'do_autosave')
+    addon = AutosaveForPlugin(None)
     addon.do_autosave.assert_not_called()
     addon.interval = 100
     addon.enabled = enabled
